Creating Things
You weren't super far of, it's just that discord can be fucking stupid sometimes. This should work for you, but from now on if you're going to make things, make sure to have a way to stop it using keys. the squiggly key will toggle it so you can type using the key, and escape will close the script. You also needed to make it send the message you want typed out, so as long as you hold down the a button it will continue to send. If you let go of the a button then it'll stop.